Rename suggest->add, remove url matching
